# Drop `HAS_TEX` for faster startup time