--- a/gwpy/plot/tex.py
+++ b/gwpy/plot/tex.py
@@ -21,6 +21,7 @@
 """
 
 import re
+from functools import lru_cache
 from shutil import which
 
 __author__ = "Duncan Macleod <duncan.macleod@ligo.org>"
@@ -44,6 +45,7 @@ def _test_usetex():
     pyplot.close(fig)
 
 
+@lru_cache(maxsize=None)
 def has_tex():
     """Returns whether tex is installed on this system
 
@@ -68,8 +70,6 @@ def has_tex():
 
     return True
 
-
-HAS_TEX = has_tex()
 
 # -- tex formatting -----------------------------------------------------------