@francois-rozet 2023-11-23
@duncanmmacleod I applied your patch!